Reorganize the serde-as code
Follow a code structure similar to serde itself.
Introduce two modules `de` and `ser`.
Put the trait implementations into `impls.rs`.
Have a dedicated `tests.rs`.
